:sparkles: Add SentimentNetwork, pre_process_data
populate review_vocab with all of the words in the given reviews.
using "split(' ')" instead of "split()".
populate label_vocab with all of the words in the given labels.
populate self.word2index with indices for all the words in self.review_vocab
do the same thing you did for self.word2index and self.review_vocab,  but for self.label2index and self.label_vocab instead.